Containerizing ingestion process
